+++ Sources/SwiftyTimer.swift
@@ -36,6 +36,7 @@ extension NSTimer {
             block()
         }
     }
+    
     /// Create a timer that will call `block` repeatedly in specified time intervals.
     ///
     /// - Note: The timer won't fire until it's scheduled on the run loop.
@@ -47,7 +48,21 @@ extension NSTimer {
             block()
         }
     }
-
+    
+    /// Create a timer that will call `block` repeatedly in specified time intervals.
+    /// (This variant also passes the timer instance to the block)
+    ///
+    /// - Note: The timer won't fire until it's scheduled on the run loop.
+    ///         Use `NSTimer.after` to create and schedule a timer in one step.
+    /// - Note: The `new` class function is a workaround for a crashing bug when using convenience initializers (rdar://18720947)
+    
+    @nonobjc public class func new(every interval: NSTimeInterval, _ block: NSTimer -> Void) -> NSTimer {
+        var timer: NSTimer!
+        timer = CFRunLoopTimerCreateWithHandler(kCFAllocatorDefault, CFAbsoluteTimeGetCurrent() + interval, interval, 0, 0) { _ in
+            block(timer)
+        }
+        return timer
+    }
 
     /// Create and schedule a timer that will call `block` once after the specified time.
     
@@ -60,6 +75,15 @@ extension NSTimer {
     /// Create and schedule a timer that will call `block` repeatedly in specified time intervals.
     
     public class func every(interval: NSTimeInterval, _ block: () -> Void) -> NSTimer {
+        let timer = NSTimer.new(every: interval, block)
+        timer.start()
+        return timer
+    }
+    
+    /// Create and schedule a timer that will call `block` repeatedly in specified time intervals.
+    /// (This variant also passes the timer instance to the block)
+    
+    @nonobjc public class func every(interval: NSTimeInterval, _ block: NSTimer -> Void) -> NSTimer {
         let timer = NSTimer.new(every: interval, block)
         timer.start()
         return timer